services/user-service/requirements.txt
fastapi==0.129.0
asyncpg
pydantic==2.12.5
python-dotenv==1.2.1
uvicorn==0.40.0
uuid-utils
sqlalchemy[asyncio]
